Reorder methods & extract callbacks into separate files
