+++ include/Controller.h
@@ -12,19 +12,14 @@ class Controller {
   explicit Controller(NimBLEAddress address);
   explicit Controller(const std::string& address);
   ~Controller() = default;
-
   bool begin(bool deleteBonds = false);
-
   bool isConnected() const;
   void onConnect(const OnConnect& callback);
   void onDisconnect(const OnDisconnect& callback);
-
   void readControls(ControlsEvent& event) const;
   void onControlsUpdate(const OnControlsUpdate& callback);
-
   void readBattery(BatteryEvent& event) const;
   void onBatteryUpdate(const OnBatteryUpdate& callback);
-
   void writeVibrations(const VibrationsCommand& cmd) const;
 
  private:
