+++ include/ControllerInternal.h
@@ -7,37 +7,30 @@
 #include "IncomingSignal.h"
 #include "OutgoingSignal.h"
 
-typedef std::function<void(ControlsEvent& e)> OnControlsUpdate;
-typedef std::function<void(BatteryEvent& e)> OnBatteryUpdate;
-typedef std::function<void(NimBLEAddress a)> OnConnect;
-typedef std::function<void(NimBLEAddress a)> OnDisconnect;
+using OnControlsUpdate = std::function<void(ControlsEvent& e)>;
+using OnBatteryUpdate = std::function<void(BatteryEvent& e)>;
+using OnConnect = std::function<void(NimBLEAddress a)>;
+using OnDisconnect = std::function<void(NimBLEAddress a)>;
 
 class ControllerInternal {
-public:
+ public:
   explicit ControllerInternal(NimBLEAddress allowedAddress);
   ~ControllerInternal() = default;
-
+  bool init(ControllerConfig& config);
+  bool deinit(bool disconnected);
   bool isInitialized() const;
-
   NimBLEAddress getAddress() const;
   void setAddress(NimBLEAddress address);
-
+  NimBLEAddress getAllowedAddress() const;
   NimBLEAddress getLastAddress() const;
   void setLastAddress(NimBLEAddress address);
-
-  NimBLEAddress getAllowedAddress() const;
-
   void onConnect(const OnConnect& callback);
   void onDisconnect(const OnDisconnect& callback);
-
   ControlsSignal& getControls();
   BatterySignal& getBattery();
   VibrationsSignal& getVibrations();
 
-  bool init(ControllerConfig& config);
-  bool deinit(bool disconnected);
-
-private:
+ private:
   bool _initialized;
   NimBLEAddress _address;
   NimBLEAddress _allowedAddress;
