+++ include/IncomingSignal.h
@@ -2,11 +2,9 @@
 
 #include <NimBLEDevice.h>
 #include <functional>
-#include "logger.h"
-#include "SignalCoder.h"
-#include "SignalConfig.h"
-#include "ControlsEvent.h"
 #include "BatteryEvent.h"
+#include "ControlsEvent.h"
+#include "IncomingSignalConfig.h"
 
 template <typename T>
 using OnUpdate = std::function<void(T& value)>;
@@ -17,23 +15,20 @@ class IncomingSignal {
   struct Store {
     T event;
   };
-
   IncomingSignal();
   ~IncomingSignal() = default;
-
-  void read(T& out);
-  bool isInitialized() const;
-  void onUpdate(const OnUpdate<T>& onUpdate);
-
   bool init(NimBLEAddress address, IncomingSignalConfig<T>& config);
   bool deinit(bool disconnected);
+  bool isInitialized() const;
+  void read(T& out);
+  void onUpdate(const OnUpdate<T>& onUpdate);
 
  private:
   static void _callConsumerFn(void* pvParameters);
+  void _handleNotify(NimBLERemoteCharacteristic* pChar, uint8_t* pData, size_t length, bool isNotify);
   bool _initialized;
   OnUpdate<T> _onUpdate;
   bool _onUpdateSet;
-  void _handleNotify(NimBLERemoteCharacteristic* pChar, uint8_t* pData, size_t length, bool isNotify);
   SignalDecoder<T> _decoder;
   NimBLEAddress _address;
   NimBLERemoteCharacteristic* _pChar;
